requirements.txt
flask>=3.0
pyyaml>=6.0
anthropic>=0.40
twilio>=9.0
python-dotenv>=1.0
gunicorn>=22.0
websockets>=12.0
pytest>=8.0
